Refactor pty code.
